Bug fix - the texts inside the UpdatesActivity got distorted while scrolling and expending views.
Root cause - The update text wrapping around the timebox was done during times it should not have been done.
Solution - Changed the timebox size to be pre-calculated, and simplified the code to always wrap the content when binding the view (instead of waiting for the timebox onMesure).

Also added expand/collapse animations when clicking more info