Bug fix - if the user facebook token expired he would see the logout button.
Root cause - When we couldn't find the token we assumed the user never logged in the past, however, this is not always the case (which caused the user to see both the cache and the logout button if his token expired).
Solution - Now using the cache existence to know if the user logged in the past, and if he did perform silent login.