Fade in images in event view and fix bug - shadow below toolbar disappeared
It disappeared when removing the default background.
For event activity, by default the background is white so the transparent
white background isn't visible when entering the screen (only the text is
visible), and this background is removed after calculating the color from
the image and fading in the image with the calculated color.
If we can't get a color from the image, the default activity background
color is used.
Using a white background color also causes the activity to fade in from
white instead of black.

Also fix fire emblem event image (remove white pixels) and update events
json.