Bug fix - Occasionally clicking a programme filter updated the UI in delay
Root cause: When clicking a programme filter, the filtering was done on a background thread using AsyncTask.execute(). By default, AsyncTask runs using a single background thread. In case clicking the filter was done in parallel to say refresh (which was also done on a background thread), this caused the filter to wait for refresh to complete.
Fix: Changed all AsyncTasks to run using a thread pool instead of a single thread, which was our original assumption on the classes behavior.