Add notification and status of sold items for second hand
Status at the top is for all open forms whose all items are sold (which
forms and the total price). This is also what the notification says in
case there are new forms with this status.
Status for each form is for all sold items in the form (number of items
and total price).

+ Make list headers sticky because forms can have up to 12 items now which
  requires scrolling
+ Add html decoding for item description
+ Add some log messages