Change event notifications format in user input
Instead of saving the notification time we save what the user selected -
the time diff between the relevant event time and the notification time.
Then, if event time is updated, we don't have to calculate the correct
time. This prevents using the incorrect time in case the cache was updated
not as part of the model refresher.