Fix feedback sending
The response of Google Forms has changed and it's no longer possible
to check for success according to the response body, so we only rely
on the response code now. This works fine for new forms but not old forms.

Also moved handling of closed form to the base class so we can know if
feedback forms are closed for some reason (instead of getting an NPE).