Support multiple time zones - common code
Amai parsers for events and updates must to be updated separetly right
after this change is merged, and the events cache should be updated.

All dates in-memory are local time. All persisted dates (in file system
and preferences) are in UTC. All dates parsed from server responses are
converted to local time. The convertion to UTC happens when writing to the
persistency, and the convertion back to the local time zone happens when
readig from persistency. Dates/times that should be in the convention time
zeon (event feedback) are converted when relevant.
Event start and end dates can be different depending on the time zone,
including the number of days in the convention. They are now calculated
according to the event days.