Fix cut off event search text
The search text must be below the title because it can be 2 lines,
which adds some margin between them, and this is not defined for the
hall/lecture texts (bottom layout).
The fix is to make the search text slightly smaller and the event view
height slighly bigger so the search text is readable (although it's still
slightly cut off).