bug fix - FAB flickers when snackbar is opened part 2
Consider the previous position of the FAB when fixing the position.
In some cases, the flicker happens when the snackbar is visible, so always
fix the position the first time.
This doesn't fix 100% of the cases but almost all of it (sometimes
the flicker happens the first time after scrolling up and down). In these
cases it seems the behavior is not called before the wrong position is
displayed.