Amai - adapt to timezone support
We don't actually support different timezones. The app dates are displayed
in Israel timezone and we convert the current time to it.
This is done because amai cons are 1 day, and we don't have a design for
multiple days, so we can't show the user's local time (since it might be
different days). Since these cons are not virtual, people in actual other
time zones can't attend anyway.