Update target sdk to 33 and handle notifications permissions
Applications require permissions to show notifications from API version 33
even without the target sdk. Setting the target sdk to 33 allows to
control when the popup is displayed.

In home screen, if the app doesn't have permissions yet to display
notifications, and it's the first time the user opened the app or the
system determines an explanation should be showed for the permissions
(this happens if the user didn't give permissions before), we show the same
popup we used to have with more descriptive text and then request the
permissions. Otherwise we only ask for the permissions without the popup.