Add support for ongoing events
Ongoing events are events which you can join at any time, which usually
last several hours or a whole day. Since they are free to join (and
sometimes leave), they don't have conflicts, aren't treated as busy time
and don't havve a start time notification. They are also not displayed
in the home screen as favorite events if there are other favorite events
are the same time.
They have a label which marks them as ongoing so the user can tell why
they are different.