Change notification settings behavior
Before, the settings affected the default notifications set when adding
events to favorite (whether the notifications were defined or not).

Now, it will always set the default notifications, but will not show them
if the settings are disabled. A message was added to the notifications
settings in the event screen in case a configured notification is disabled.

The change is because when the setting is disabled, it's unexpected to get
notficiations even if the events were added to the favorites before
disabling it. It is also unexpected not to get notifications if the setting
is enabled, even if you added the event to favorites while it was disabled.