Support map highlight color per stand location/area and blend modes
For the blend modes to work, Android 10 (API 29) or above is required,
so we don't highlight for lower versions (~1% of users in the last con).

In the future this change can allow to paint stand highlights in other
shapes besides a rectangle.

Since the highlight color is now per stand, it's no longer an attribute.
The standsMapHighlight attribute was removed.