Add padding for display cutouts (requires amai adaptation)
Required adaptation: in amai-specific activities, change the call to
registerApplyInsets (the start and end insets are separated now) and
add side padding to the relevant views since it was removed from the
content view. Make sure no text or buttons overlaps the display cutouts
and navigation buttons in landscape mode.

In immersive mode, add a padding for the display cutouts but allow to
draw under them with clipToPadding="false".