Sff - fix bug in second hand
When the second tab (sell) is selected and the screen orientation changes,
the tab changes don't work at first, until selecting the other one and
then that one again.
It happened because we didn't set last the selected tab (we always
set the first tab) but the UI still displayed it.

Also fix depracated API usage.