Fix moovit integration
Since API level 30, apps can't get info about some other apps unless
specified in the manifest. This makes the check if moovit is installed
always return false, so the browser was always opened.
Since resolveActivity doesn't always work as expected, the way to open
deep links (and web links) is to call startActivity and catch the
exception (which we already did anyway).