Remove parallax from event activity and fix scroll bug
The bug is that when scrolling up touching the event description, it
doesn't support nested scrolling (it doesn't scroll away the appbar).
It happens because the movement method handles scrolling.
Using a NestedScrollView instead of ScrollView fixes the issue.

I also changed ParallaxScrollView to be a NestedScrollView, which solved
it, but we don't use the parallax feature anymore so I removed it from
the event activity.