Fix edge to edge scrolling in text views with links (requires amai adaptation)
Relevant for: event activity, about activity.

This reverts the NestedScrollView change in event activity.

The bug is that inside a scroll view, when the last element is a text
view which has a LinkMovementMethod, and you drag the scroll view up
from that text view, the scroll doesn't include the padding of the
scroll view so it doesn't scroll all the way down.

The fix is to add another view which will get the bottom padding instead
of adding it to the scroll view.

Amai adaptation: in all amai-specific activities that have a scroll view
which contains a text view with a LinkMovementMethod, add this view at
the bottom (best if there is no background to it):
<View
  android:layout_width="match_parent"
  android:layout_height="0dp"
  android:id="@+id/scroll_view_bottom_padding"/>

In the edge to edge handling in the activity, send NONE as the bottom
inset for the scroll view, and add this code:
// The bottom padding must be applied to a view inside the scroll view, otherwise the scrolling doesn't include the padding when scrolled from the last
// textview and it has a LinkMovementMethod
Views.registerApplyInsets(Views.InsetType.NONE, Views.InsetType.PADDING, Views.InsetType.NONE, Views.InsetType.NONE, findViewById(R.id.scroll_view_bottom_padding));