Support passing change on keyboard open (requires amai adaptation)
Before edge to edge, the adjustResize setting for softInputMode in the
manifest took care of changing the activity size when the keyboard was
opened. In edge to edge, we have to add ime to the padding adjustment
to support it.

Amai adaptation: registerApplyInsets parameter for ime was added.
In activities where no resize is necessary, set the new parameter to false.
Where it's required, set the new parameter to true.

Also -
1. Support passing a context instead of activity to
   hideKeyboardOnClickOutsideEditText (for fragments)
2. Decrease item send animation duration from 5 seconds to 3 seconds in
   send all in feedback activity